requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
Faker==40.1.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
psycopg2==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.4
